refactor(control): create components for control informations page
