+++ web/controller/components/details/ControllerControlDayPageInfo.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+export function ControllerControlDayPageInfo() {
+  return <></>;
+}
